# api.yaml
openapi: 3.0.1
info:
  title: Product API
  version: 1.0.0
components:
  schemas:
    product:
      type: object
      properties:
        id:
          type: integer
          minimum: 0
          example: 12
        name:
          type: string
          example: banana
        stock:
          type: integer
          minimum: 0
        price:
          type: number
          minimum: 0
        updated:
          type: string
          format: 'date-time'
      example:
        id: 1
        name: banana
        stock: 20
        price: 0.23
        updated: 2014-01-02T00:00:00.000Z
paths:
  /products:
    put:
      summary: Replaces products.
      operationId: ReplaceProducts
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/product'
          description: OK
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                type: string
    get:
      summary: Returns a list of products.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/product'
          description: OK
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                type: string
  /products/{id}:
    patch:
      summary: Modifies a product.
      parameters:
        - required: true
          in: path
          name: id
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: OK
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                type: string
  /products/search:
    get:
      summary: Returns a list of products.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/product'
          description: OK
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                type: string